+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 2025-11-02 : 0.4.0-beta.9 (prerelease)  
 Update: Round of corrections, updates; Testing in hosting environment
 
+- Added auth.providers.roles. config namespace for defining Roles
 - Correct handling of the lowest role when seeding factory users
 - Correct how model builder is called in the user index query()
 - Correct issue identifying user myself in the validation rules